Potential fix for code scanning alert no. 29: Multiplication result converted to larger type

+++ test/testyuv.c
@@ -525,8 +525,8 @@ static bool create_textures(SDL_Renderer *renderer, SDL_Surface *original, SDL_P
         const Uint8 *src_plane1 = src_plane0 + Yrows * src_Ypitch;
         const int Ypitch = pitch + 37;
         const int UVpitch = ((Ypitch + 1) / 2) * 2;
-        Uint8 *plane0 = (Uint8 *)SDL_calloc(1, Yrows * Ypitch);
-        Uint8 *plane1 = (Uint8 *)SDL_calloc(1, UVrows * UVpitch);
+        Uint8 *plane0 = (Uint8 *)SDL_calloc(1, (size_t)Yrows * (size_t)Ypitch);
+        Uint8 *plane1 = (Uint8 *)SDL_calloc(1, (size_t)UVrows * (size_t)UVpitch);
         int row;
         const Uint8 *src;
         Uint8 *dst;
